digital_attendance/requirements.txt
streamlit>=1.30.0
streamlit-webrtc>=0.45.0
opencv-python-headless>=4.8.0
pillow>=10.0.0
pandas>=2.1.0
qrcode>=7.4.2
av>=10.0.0
numpy>=1.26.0
